pyproject.toml
[project]
name = "image-video-url-maker-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cloudinary>=1.44.0",
    "fastapi>=0.115.12",
    "pillow>=11.2.1",
    "python-dotenv>=1.1.0",
    "pytz>=2025.2",
    "requests>=2.32.4",
    "uv>=0.7.13",
    "uvicorn>=0.34.3",
]
